requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
uvicorn[standard]
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
starlette==0.27.0
itsdangerous
pytest==7.4.2
chromadb==0.4.22
google-generativeai==0.3.2
gtts==2.4.0
pyttsx3==2.90
python-dotenv==1.0.0
numpy<2.0.0

